Move User models to models.py
